Refactor database_creation.py Refactor discord-script.py
