+++ scraping_google_images.py
@@ -8,7 +8,6 @@ from selenium.webdriver.chrome.service import Service as ChromeService
 from webdriver_manager.chrome import ChromeDriverManager
 from selenium.webdriver.common.by import By
 
-
 # Driver for ChromeDriver
 service = ChromeService(ChromeDriverManager().install())
 wd = webdriver.Chrome(service=service)
